requirements.txt
customtkinter
pillow
darkdetect